updates to file structure, welcome message, and added more schools to data json
