+++ app/Welcome.js
@@ -17,25 +17,25 @@ export default  function Welcome({options, setOptions}){
   const handleCheck = (e) => {
     setWelcomeCheck(e.target.checked)
   }
-  console.log(welcomeCheck)
+  console.log(open)
   return (
     <div>
       <Modal
-        keepMounted
         open={open}
         onClose={handleClose}
-        aria-labelledby="keep-mounted-modal-title"
-        aria-describedby="keep-mounted-modal-description"
       >
         <Box className="absolute top-1/2 left-1/2 transform -translate-x-1/2 -translate-y-1/2 w-[700px] bg-black border-4 border-white border-opacity-90 bg-opacity-80 p-6 rounded">
-          <Typography className='text-center font-bold text-2xl text-white' id="keep-mounted-modal-title">
-            Welcome to the college football conference map
+          <Typography className='text-center font-bold text-xl text-white' id="keep-mounted-modal-title">
+            WECLOME TO THE COLLEGE FOOTBALL CONFERENCE HISTORY MAP
           </Typography>
           <Typography className="text-center mt-5 font-normal text-white">
             With the recent chaos of conference realignment, it is a good time to look back on how we got here.
           </Typography>
           <Typography className="text-center mt-5 font-normal text-white">
-            This conference map will start in YEAR before the creation of the NCAA. It will then follow along the Division I Conferences, until the 1978 Division I split and continue following the Division IA Conferences, which would later be known as the Football Bowl Subdivision. 
+            This conference map will start in 1894 before the creation of the NCAA. It will then follow along the Division I Conferences, until the 1978 Division I split and continue following the Division IA Conferences, which would later be known as the Football Bowl Subdivision. 
+          </Typography>
+          <Typography className="text-center mt-5 font-normal text-white">
+            You can use the left and right arrows on your keyboard to change years.
           </Typography>
           <Typography className="text-center mt-5 font-normal text-white">
             Checkout the settings menu for filtering options and display settings.
@@ -68,7 +68,6 @@ export default  function Welcome({options, setOptions}){
               <Button className='text-black font-bold bg-white border-white hover:bg-black hover:text-white hover:border-white' variant='outlined' onClick={handleClose}>Close</Button>
             </div>
           </div>
-
         </Box>
       </Modal>
     </div>
