+++ app/data/cfbHistory.json
@@ -1,0 +1,11 @@
+{
+  "history": [
+    {
+      "year": 1906,
+      "events": [
+        "The Intercollegiate Athletic Association of the United States (IAAUS) was established on March 31, 1906 and would later be known as the National Collegiate Athletic Association (NCAA).",
+        "College Football under the IAAUS began as one division."
+      ]
+    }
+  ]
+}
