+++ app/functions/getConferences.js
