+++ next.config.js
@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   images: {
-    domains: ['upload.wikimedia.org', 'collegepolltracker.com','1000logos.net', 'www.pulf.org', 'i.pinimg.com'], // Add the relevant domains here
+    domains: ['upload.wikimedia.org', 'collegepolltracker.com','1000logos.net', 'www.pulf.org', 'i.pinimg.com','static.wikia.nocookie.net'], // Add the relevant domains here
   },
 };
